requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
